These were made during Revision development:
Added AnimTweenBlendTrigger to let Pawn's blend an animation with something they are already playing.
Added ElevatorTrigger to send EelevatorMovers to different keyframes depending on what button triggered it.
Added FlaggedActorTrigger that sets a flag with the name of the actor that triggered it.
Added LOSHateTrigger that makes pawns that see the triggerer hate them.
Added OrderPawnTrigger to give orders to pawns that trigger it.
Added UnlitTrigger to toggle Actors bUnlit property.